+++ util/util.go
@@ -1,0 +1,2 @@
+// Package util adds additional utilities.
+package util